KernelInterface: dummy_input_params for kernel interfaces
Driven by backward compatibility, in order to keep the
original kernel interface still, the number of arguments
should not be changed when we are intergrating the semantics
of core-tagging object retyping operation through UT_RT.

When we are trying to hold the old interface for system
designers, it's also preferable to consider not introducing
new kernel object type for the Per-Core Objects, which
should be treated as a property for taggable objects.
The actual need for implementing Per-Core Objects is taken
care of by using a dummy param. The dummy params are not
specified in the kernel interface for original interface
semantics, instead, their places will be set to nil in
the message buffer, hence dummy.

This commit is a scratch implementation of dummy input
for system interfaces.

Signed-off-by: Zhu Guangtao <guangtaozhu@outlook.com>